# docs/api-spec/loyalty-rewards-service-openapi.yaml
openapi: 3.0.3
info:
  contact:
    email: loyalty@probestack.io
    name: Growth Engineering Team
  description: |
    Manages loyalty point balances, tier progression, and reward redemptions.
    This API is used as the source specification to automatically generate
    an API Gateway proxy (Kong or Apigee) with routing, authentication,
    rate limiting, and caching policies.
  title: Loyalty Rewards Service
  version: 1.0.0
  x-proxy-config:
    upstream-url: https://loyalty-service.probestack.io
    service-name: loyalty-rewards-gateway
    rate-limit: 250/minute
    cors-enabled: true
    retries: 2
servers:
- description: Production Gateway
  url: https://gateway.probestack.io/loyalty/v1
- description: Local Gateway
  url: http://localhost:8000/loyalty/v1
security:
- ApiKeyAuth: []
...
paths:
...
  /accounts/{memberId}/transactions:
    get:
      description: "Returns the earn and burn history for a member, most recent first."
      operationId: listPointTransactions
      parameters:
      - explode: false
        in: path
        name: memberId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      - description: Filter by transaction type
        explode: true
        in: query
        name: type
        required: false
        schema:
          default: ALL
          enum:
          - EARN
          - BURN
          - EXPIRY
          - ADJUSTMENT
          - ALL
          type: string
        style: form
      responses:
        "200":
          content:
            application/json:
              schema:
                items:
                  $ref: "#/components/schemas/PointTransaction"
                type: array
          description: Transactions fetched successfully
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid API Key.
        "404":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Resource not found.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: List point transactions
      tags:
      - Transactions
      x-proxy-route:
        methods:
        - GET
        plugins:
        - name: cors
        - name: rate-limiting
          config:
            minute: 250
      x-accepts: application/json
      x-tags:
      - tag: Transactions
    post:
      description: Credits or debits points against the member's balance. Burns are
        rejected if the balance is insufficient.
      operationId: createPointTransaction
      parameters:
      - explode: false
        in: path
        name: memberId
        required: true
        schema:
          format: uuid
          type: string
        style: simple
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateTransactionRequest"
        required: true
      responses:
        "201":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PointTransaction"
          description: Transaction recorded successfully
        "400":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Invalid request payload or parameters.
        "401":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Missing or invalid API Key.
        "500":
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
          description: Internal server error.
      summary: Record a point transaction
      tags:
      - Transactions
      x-proxy-route:
        methods:
        - POST
        plugins:
        - name: cors
        - name: rate-limiting
          config:
            minute: 60
      x-content-type: application/json
      x-accepts: application/json
      x-tags:
      - tag: Transactions
...
components:
...
  schemas:
...
    PointTransaction:
      example:
        occurredAt: 2000-01-23T04:56:07.000+00:00
        balanceAfter: 12450
        id: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        source: ORDER:ORD-2026-0001
        type: EARN
        memberId: 046b6c7f-0b8a-43b9-b35d-6489e6daee91
        points: 350
      properties:
        id:
          format: uuid
          type: string
        memberId:
          format: uuid
          type: string
        type:
          enum:
          - EARN
          - BURN
          - EXPIRY
          - ADJUSTMENT
          example: EARN
          type: string
        points:
          description: "Positive for earn, negative for burn or expiry"
          example: 350
          type: integer
        source:
          example: ORDER:ORD-2026-0001
          type: string
        balanceAfter:
          example: 12450
          type: integer
        occurredAt:
          format: date-time
          type: string
      type: object
    CreateTransactionRequest:
      example:
        note: Points for August order
        source: ORDER:ORD-2026-0001
        type: EARN
        points: 350
      properties:
        type:
          enum:
          - EARN
          - BURN
          - ADJUSTMENT
          example: EARN
          type: string
        points:
          example: 350
          type: integer
        source:
          example: ORDER:ORD-2026-0001
          type: string
        note:
          example: Points for August order
          type: string
      required:
      - points
      - type
      type: object
...
    ErrorResponse:
      properties:
        status:
          example: 400
          type: integer
        errorCode:
          example: INSUFFICIENT_POINTS
          type: string
        message:
          example: Member does not have enough points for this reward
          type: string
        timestamp:
          format: date-time
          type: string
        path:
          example: /loyalty/v1/rewards
          type: string
      required:
      - message
      - status
      type: object
  securitySchemes:
    ApiKeyAuth:
      description: API Key required for all gateway requests.
      in: header
      name: X-API-Key
      type: apiKey