# src/main/java/com/probestack/forgestudio/design/api/TransactionsApi.java
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.2.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package com.probestack.forgestudio.design.api;

import com.probestack.forgestudio.design.model.CreateTransactionRequest;
import com.probestack.forgestudio.design.model.ErrorResponse;
import com.probestack.forgestudio.design.model.PointTransaction;
import java.util.UUID;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-08-13T10:36:55.986768776Z[GMT]")@Validated
@Tag(name = "Transactions", description = "Earn and burn point transactions")
public interface TransactionsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    /**
     * POST /accounts/{memberId}/transactions : Record a point transaction
     * Credits or debits points against the member&#39;s balance. Burns are rejected if the balance is insufficient.
     *
     * @param memberId  (required)
     * @param createTransactionRequest  (required)
     * @return Transaction recorded successfully (status code 201)
     *         or Invalid request payload or parameters. (status code 400)
     *         or Missing or invalid API Key. (status code 401)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "createPointTransaction",
        summary = "Record a point transaction",
        description = "Credits or debits points against the member's balance. Burns are rejected if the balance is insufficient.",
        tags = { "Transactions" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Transaction recorded successfully", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = PointTransaction.class))
            }),
            @ApiResponse(responseCode = "400", description = "Invalid request payload or parameters.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Missing or invalid API Key.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = "/accounts/{memberId}/transactions",
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    
    default ResponseEntity<PointTransaction> createPointTransaction(
        @Parameter(name = "memberId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("memberId") UUID memberId,
        @Parameter(name = "CreateTransactionRequest", description = "", required = true) @Valid @RequestBody CreateTransactionRequest createTransactionRequest
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "{ \"occurredAt\" : \"2000-01-23T04:56:07.000+00:00\", \"balanceAfter\" : 12450, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"source\" : \"ORDER:ORD-2026-0001\", \"type\" : \"EARN\", \"memberId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"points\" : 350 }";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }


    /**
     * GET /accounts/{memberId}/transactions : List point transactions
     * Returns the earn and burn history for a member, most recent first.
     *
     * @param memberId  (required)
     * @param type Filter by transaction type (optional, default to ALL)
     * @return Transactions fetched successfully (status code 200)
     *         or Missing or invalid API Key. (status code 401)
     *         or Resource not found. (status code 404)
     *         or Internal server error. (status code 500)
     */
    @Operation(
        operationId = "listPointTransactions",
        summary = "List point transactions",
        description = "Returns the earn and burn history for a member, most recent first.",
        tags = { "Transactions" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Transactions fetched successfully", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = PointTransaction.class)))
            }),
            @ApiResponse(responseCode = "401", description = "Missing or invalid API Key.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Resource not found.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "500", description = "Internal server error.", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "ApiKeyAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = "/accounts/{memberId}/transactions",
        produces = { "application/json" }
    )
    
    default ResponseEntity<List<PointTransaction>> listPointTransactions(
        @Parameter(name = "memberId", description = "", required = true, in = ParameterIn.PATH) @PathVariable("memberId") UUID memberId,
        @Parameter(name = "type", description = "Filter by transaction type", in = ParameterIn.QUERY) @Valid @RequestParam(value = "type", required = false, defaultValue = "ALL") String type
    ) {
                        getRequest().ifPresent(request -> {
                    for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                        if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                            String exampleString = "[ { \"occurredAt\" : \"2000-01-23T04:56:07.000+00:00\", \"balanceAfter\" : 12450, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"source\" : \"ORDER:ORD-2026-0001\", \"type\" : \"EARN\", \"memberId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"points\" : 350 }, { \"occurredAt\" : \"2000-01-23T04:56:07.000+00:00\", \"balanceAfter\" : 12450, \"id\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"source\" : \"ORDER:ORD-2026-0001\", \"type\" : \"EARN\", \"memberId\" : \"046b6c7f-0b8a-43b9-b35d-6489e6daee91\", \"points\" : 350 } ]";
                            ApiUtil.setExampleResponse(request, "application/json", exampleString);
                            break;
                        }
                    }
                });
                return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);
                                                            }

}